filter events don't call callbacks on every BEV_OK
It was misunderstood that filtered bufferevents would invoke the
registered read_cb for every BEV_OK returned from the filter. This is
not the case.

filter_pbrpc_messgaes ensures that read_cb is called when one or more
messages have arrived in full. So, svc_read_cb and clnt_read_cb must
drain all the messages that were passed on.

Signed-off-by: Krishnan Parthasarathi <kparthas@redhat.com>